- fixed the code based on the requirements
